Update BaseScene.unity
I added a script for controlling a ship. currently I need to work on how it feels to move and the controls for it.

I created a script to spawn a field of asteroids inside a cube of size, size. It also checks that the asteroids spawned won't be overlapped.

I think I'm going to start working on the part of the game where the player can exit the ship for repairs or gathering materials.

That will start with me changing the ship controls to be wasd for moving forward backward left right then Q and E for rotating.

Then I'll just focus on making another controller for the player themselves afterwards.