Update ShipController.cs
I added something to clamp the speed of the ship. I also added some parameters for the rotation speed.

I also started implementing the damage and think I need to set up the canvas to display the health.

Then after that I can implement the stuff for needing to repair the ship with scrap.