Update PlayerController.cs
I started getting aligning the player with the ship working properly but I didn't realize I've spent 2 hours working on it until now.

I'm close to getting it correct but I need to work more on that aspect.

I have other stuff to do so I'm going to go work on that then come back to this.