Repair minigame fully implemented.
I have the repair minigame fully implemented such that damage affects the hull integrity and when you repair the hull integrity goes back up to the correct health.

10 spots of damage = death.
10 health per damage decal so at 1 damage decal health is 90. Repairing simply adds the health back to normal.

Also, I need to code it so that scrap can be collected and that is what you have to use to repair the ship.