Update BaseScene.unity (Implement invisblewalls)
I implemented a proximity shader that reveals an invisible wall that says "WARNING!!! OUT OF BOUNDS" when the player is close to it.

I also implemented the earth far in the distance. But, far enough to know you're probably near the same distance as the moon.

Next is to implement the oxygen system for the player.

Then I'll implement the scrap, and the gems. (Collectables).

Then I'll hopefully be able to draw up some prefabs of derelict stations and maybe a ship or two that's the same as yours that contains random amounts of scrap and contains a note about the lore.

I think I'll just quickly make something that spawns the different asteroids and always spawns the red gem at a farther distance to make sure it scares the player and kills them after they begin exploring.