Added UI stuff and Scrap stuff.
UI additions for scrap added a 3D model for it. Also added something that fixed how you get into the ship where you have to look at the ship console and press F and then you start controlling the ship. It also teleports you to the right position to control the ship.