Implemented credits and ending condition/screen.
I added a GameManager that stores the endCondition so when you return to the title screen it'll display it with the title.

I also added a script that hides the quit button in WebGL only.

Made it so that when you get killed by the angler you get the actual game's ending.